+++ EveLib.EveCentral/Models/EveCentralMarketStatResponse.cs
@@ -13,7 +13,7 @@ namespace eZet.EveLib.Modules.Models {
 
     public class EveCentralMarketStatItem {
         [XmlAttribute("id")]
-        public long TypeId { get; set; }
+        public int TypeId { get; set; }
 
         [XmlElement("buy")]
         public EveCentralMarketStatOrderData BuyOrders { get; set; }
@@ -30,21 +30,21 @@ namespace eZet.EveLib.Modules.Models {
         public long Volume { get; set; }
 
         [XmlElement("avg")]
-        public double Average { get; set; }
+        public decimal Average { get; set; }
 
         [XmlElement("max")]
-        public double Max { get; set; }
+        public decimal Max { get; set; }
 
         [XmlElement("min")]
-        public double Min { get; set; }
+        public decimal Min { get; set; }
 
         [XmlElement("stddev")]
         public double StdDev { get; set; }
 
         [XmlElement("median")]
-        public double Median { get; set; }
+        public decimal Median { get; set; }
 
         [XmlElement("percentile")]
-        public double Percentile { get; set; }
+        public decimal Percentile { get; set; }
     }
 }
